install drywall-compile in ./node_modules/.bin
Add a script to build the tools.
The tool should be run from root of repo like this:
$ ./node_modules/.bin/drywall-compile

The compells drywall to download assets needed via bower, then builds
the assets into the public subdirectory using grunt.
I was unable to find a way to get grunt to work during npm install
process, or as part of the "postinstall" script.  However, running the
script once it has been installed from root of repo does seem to work.

There are warnings on my system for things like "grunt-contrib-copy"
not found, is it installed on the system, so I've left in some
debugging I attempted to try to get these warnings to go away.
Despite the warnings, none of my attempts (shown in comments) worked
to eliminate the messages, and yet the build seems to produce correct
results.